backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
psycopg2-binary==2.9.9
python-dotenv==0.19.0